pyproject.toml
[tool.poetry]
name = "sw"
version = "0.1.0"
description = ""
authors = ["Aleksander Philips"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.3"
django-environ = "^0.8.1"
requests = "^2.27.1"
petl = "^1.7.8"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = "^22.3.0"
isort = "^5.10.1"
pdbpp = "^0.10.3"
Faker = "^13.3.4"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
